refactor: add SaveConfigButton view for settings
