+++ Meshtastic/Views/Settings/Config/BluetoothConfig.swift
@@ -12,7 +12,6 @@ struct BluetoothConfig: View {
 	@EnvironmentObject var bleManager: BLEManager
 	@Environment(\.dismiss) private var goBack
 	var node: NodeInfoEntity?
-	@State private var isPresentingSaveConfirm: Bool = false
 	@State var hasChanges = false
 	@State var enabled = true
 	@State var mode = 0
@@ -71,42 +70,24 @@ struct BluetoothConfig: View {
 			}
 		}
 		.disabled(self.bleManager.connectedPeripheral == nil || node?.bluetoothConfig == nil)
-		Button {
-			isPresentingSaveConfirm = true
-		} label: {
-			Label("save", systemImage: "square.and.arrow.down")
-		}
-		.disabled(bleManager.connectedPeripheral == nil || !hasChanges || shortPin)
-		.buttonStyle(.bordered)
-		.buttonBorderShape(.capsule)
-		.controlSize(.large)
-		.padding()
-		.confirmationDialog(
-			"are.you.sure",
-			isPresented: $isPresentingSaveConfirm,
-			titleVisibility: .visible
-		) {
-			let nodeName = node?.user?.longName ?? "unknown".localized
-			let buttonText = String.localizedStringWithFormat("save.config %@".localized, nodeName)
-			Button(buttonText) {
-				let connectedNode = getNodeInfo(id: bleManager.connectedPeripheral.num, context: context)
-				if connectedNode != nil {
-					var bc = Config.BluetoothConfig()
-					bc.enabled = enabled
-					bc.mode = BluetoothModes(rawValue: mode)?.protoEnumValue() ?? Config.BluetoothConfig.PairingMode.randomPin
-					bc.fixedPin = UInt32(fixedPin) ?? 123456
-					let adminMessageId =  bleManager.saveBluetoothConfig(config: bc, fromUser: connectedNode!.user!, toUser: node!.user!, adminIndex: connectedNode?.myInfo?.adminIndex ?? 0)
-					if adminMessageId > 0 {
-						// Should show a saved successfully alert once I know that to be true
-						// for now just disable the button after a successful save
-						hasChanges = false
-						goBack()
-					}
+
+		SaveConfigButton(node: node, hasChanges: $hasChanges) {
+			let connectedNode = getNodeInfo(id: bleManager.connectedPeripheral.num, context: context)
+			if connectedNode != nil {
+				var bc = Config.BluetoothConfig()
+				bc.enabled = enabled
+				bc.mode = BluetoothModes(rawValue: mode)?.protoEnumValue() ?? Config.BluetoothConfig.PairingMode.randomPin
+				bc.fixedPin = UInt32(fixedPin) ?? 123456
+				let adminMessageId =  bleManager.saveBluetoothConfig(config: bc, fromUser: connectedNode!.user!, toUser: node!.user!, adminIndex: connectedNode?.myInfo?.adminIndex ?? 0)
+				if adminMessageId > 0 {
+					// Should show a saved successfully alert once I know that to be true
+					// for now just disable the button after a successful save
+					hasChanges = false
+					goBack()
 				}
 			}
-		} message: {
-			Text("config.save.confirm")
 		}
+
 		.navigationTitle("bluetooth.config")
 		.navigationBarItems(trailing:
 			ZStack {
