+++ Meshtastic/Views/Settings/Config/Module/MQTTConfig.swift
@@ -191,48 +191,28 @@ struct MQTTConfig: View {
 			.scrollDismissesKeyboard(.interactively)
 			.disabled(self.bleManager.connectedPeripheral == nil || node?.mqttConfig == nil)
 		}
-		Button {
-			isPresentingSaveConfirm = true
-		} label: {
-			Label("save", systemImage: "square.and.arrow.down")
-		}
-		.disabled(bleManager.connectedPeripheral == nil || !hasChanges)
-		.buttonStyle(.bordered)
-		.buttonBorderShape(.capsule)
-		.controlSize(.large)
-		.padding()
-		.confirmationDialog(
-			"are.you.sure",
-			isPresented: $isPresentingSaveConfirm,
-			titleVisibility: .visible
-		) {
+
+		SaveConfigButton(node: node, hasChanges: $hasChanges) {
 			let connectedNode = getNodeInfo(id: bleManager.connectedPeripheral?.num ?? -1, context: context)
 			if connectedNode != nil {
-				let nodeName = node?.user?.longName ?? "unknown".localized
-				let buttonText = String.localizedStringWithFormat("save.config %@".localized, nodeName)
-				Button(buttonText) {
-					var mqtt = ModuleConfig.MQTTConfig()
-					mqtt.enabled = self.enabled
-					mqtt.proxyToClientEnabled = self.proxyToClientEnabled
-					mqtt.address = self.address
-					mqtt.username = self.username
-					mqtt.password = self.password
-					mqtt.root = self.root
-					mqtt.encryptionEnabled = self.encryptionEnabled
-					mqtt.jsonEnabled = self.jsonEnabled
-					mqtt.tlsEnabled = self.tlsEnabled
-					let adminMessageId =  bleManager.saveMQTTConfig(config: mqtt, fromUser: connectedNode!.user!, toUser: node!.user!, adminIndex: connectedNode?.myInfo?.adminIndex ?? 0)
-					if adminMessageId > 0 {
-						// Should show a saved successfully alert once I know that to be true
-						// for now just disable the button after a successful save
-						hasChanges = false
-						goBack()
-					}
+				var mqtt = ModuleConfig.MQTTConfig()
+				mqtt.enabled = self.enabled
+				mqtt.proxyToClientEnabled = self.proxyToClientEnabled
+				mqtt.address = self.address
+				mqtt.username = self.username
+				mqtt.password = self.password
+				mqtt.root = self.root
+				mqtt.encryptionEnabled = self.encryptionEnabled
+				mqtt.jsonEnabled = self.jsonEnabled
+				mqtt.tlsEnabled = self.tlsEnabled
+				let adminMessageId =  bleManager.saveMQTTConfig(config: mqtt, fromUser: connectedNode!.user!, toUser: node!.user!, adminIndex: connectedNode?.myInfo?.adminIndex ?? 0)
+				if adminMessageId > 0 {
+					// Should show a saved successfully alert once I know that to be true
+					// for now just disable the button after a successful save
+					hasChanges = false
+					goBack()
 				}
 			}
-		}
-		message: {
-			Text("config.save.confirm")
 		}
 		.navigationTitle("mqtt.config")
 		.navigationBarItems(trailing:
